rule_engine : cache execution_context creation
Currently, every call to the rule_engine.compute method forces the recreation of
all the functions which *may* be use by the rule.

We set up a cache that stores the context for a given rule. This cache is not
used on not yet saved rules, or in debug mode.

review http://rietveld.coopengo.com/5080001/