Coog: Skip 'all_options' field in export of 'Covered Element' model (#134)
Export of both 'all_options' and 'options' fields in the 'contract.covered_element' model leads to duplicates as they are both One2Many fields on the same target. The only difference is that 'options' has a domain. I added 'all_options' to export skips in order to avoid that.

Fix #5199