premium : Recalculate premiums at least at the given date (#265)
The calculate_prices method accepts a 'start' parameter. The behaviour of the
method is to trim all premium lines which ends after this parameter to
start - 1, then calculate all prices at the dates at which the contract premium
may change. The problem is that if no dates are found (or if the 'start' date
is not considered 'relevant' on the contract, there will be a hole in the
premiums for this contract.

We should always calculate prices at least at the 'start' date that was given
as a parameter.

Fix #5570